C9 - Added endpoint and service method for getting all documents.

+++ src/main/java/com/ld/documentarchive/service/impl/DocumentServiceImpl.java
@@ -2,14 +2,18 @@ package com.ld.documentarchive.service.impl;
 
 import com.ld.documentarchive.dto.DocumentDto;
 import com.ld.documentarchive.entity.Document;
+import com.ld.documentarchive.exception.DocumentException;
 import com.ld.documentarchive.repo.DocumentRepository;
 import com.ld.documentarchive.service.DocumentService;
+import com.ld.documentarchive.utils.Mapper;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -20,31 +24,37 @@ public class DocumentServiceImpl implements DocumentService {
 
     @Override
     public DocumentDto createDocument(DocumentDto documentDto) {
-        ModelMapper modelMapper = new ModelMapper();
-        documentRepository.save(modelMapper.map(documentDto, Document.class), Duration.ZERO);
+        documentRepository.save(Mapper.mapDocumentDTOToDocument(documentDto), Duration.ZERO);
         return documentDto;
     }
 
     @Override
+    public List<DocumentDto> readAllDocuments() {
+        List<Document> documentList = documentRepository.findAll();
+        return documentList.stream().map(Mapper::mapDocumentToDocumentDTO).collect(Collectors.toList());
+    }
+
+    @Override
     public DocumentDto readDocumentById(String documentId) {
-        ModelMapper modelMapper = new ModelMapper();
         Document document = documentRepository.findById(documentId).orElseThrow();
-        return modelMapper.map(document, DocumentDto.class);
+        return Mapper.mapDocumentToDocumentDTO(document);
     }
 
     @Override
     public DocumentDto updateDocumentById(String documentId, DocumentDto documentDto) {
-        ModelMapper modelMapper = new ModelMapper();
         Document document = documentRepository.findById(documentId).orElseThrow();
         document.setDocumentType(documentDto.getDocumentType());
         documentRepository.save(document, Duration.ZERO);
-        return modelMapper.map(document,DocumentDto.class);
+        return Mapper.mapDocumentToDocumentDTO(document);
     }
 
     @Override
     public String deleteDocumentById(String documentId) {
-        Document document = documentRepository.findById(documentId).orElseThrow();
-        documentRepository.deleteById(documentId);
-        return documentId;
+        if (documentRepository.existsById(documentId)) {
+            documentRepository.deleteById(documentId);
+            return documentId;
+        } else {
+            throw new DocumentException("Document not found.");
+        }
     }
 }
